Merging my fixes

+++ src/main/java/ca/mcgill/ecse321/fitnessplusplus/service/RegisteredUserService.java
@@ -1,9 +1,7 @@
 package ca.mcgill.ecse321.fitnessplusplus.service;
 
-import ca.mcgill.ecse321.fitnessplusplus.model.Client;
-import ca.mcgill.ecse321.fitnessplusplus.model.Instructor;
-import ca.mcgill.ecse321.fitnessplusplus.model.Owner;
-import ca.mcgill.ecse321.fitnessplusplus.model.RegisteredUser;
+import ca.mcgill.ecse321.fitnessplusplus.dto.OwnerDto;
+import ca.mcgill.ecse321.fitnessplusplus.model.*;
 
 import ca.mcgill.ecse321.fitnessplusplus.repository.ClientRepository;
 import ca.mcgill.ecse321.fitnessplusplus.repository.InstructorRepository;
@@ -13,6 +11,9 @@ import ca.mcgill.ecse321.fitnessplusplus.repository.RegisteredUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class RegisteredUserService {
@@ -47,6 +48,11 @@ public class RegisteredUserService {
 
     @Transactional
     public RegisteredUser promoteUser(RegisteredUser registeredUser) {
+
+        if (registeredUser == null) {
+            throw new IllegalArgumentException("You cannot promote a user that does not exist");
+        }
+
         if (registeredUser.getAccountRole().getClass() == Client.class) {
             // Delete client account role -> change to instructor
             clientRepository.delete((Client) registeredUser.getAccountRole());
@@ -75,4 +81,34 @@ public class RegisteredUserService {
 
         return registeredUser;
     }
+
+    @Transactional
+    public List<RegisteredUser> getAllRegisteredUser() {
+        List<RegisteredUser> list = new ArrayList<RegisteredUser>();
+
+        for (RegisteredUser r : registeredUserRepository.findAll()) {
+            list.add(r);
+        }
+
+        return list;
+    }
+
+    @Transactional
+    public RegisteredUser getRegisteredUserById(int id){return registeredUserRepository.findRegisteredUserByUserId(id);}
+
+    @Transactional
+    public Client getClientById(int id){
+        return clientRepository.findClientByroleId(id);
+    }
+
+    @Transactional
+    public Instructor getInstructorById(int id){
+        return instructorRepository.findInstructorByroleId(id);
+    }
+
+    @Transactional
+    public Owner getOwnerById(int id){
+        return ownerRepository.findInstructorByroleId(id);
+    }
+
 }
